+++ src/pyrokinetics/gk_code/tglf.py
@@ -22,7 +22,7 @@ from ..normalisation import convert_dict
 from ..numerics import Numerics
 from ..templates import gk_templates
 from ..typing import PathLike
-from .gk_input import GKInput
+from .gk_input import GKInput, GKInputFlat
 from .gk_output import (
     Coords,
     Eigenfunctions,
@@ -34,7 +34,7 @@ from .gk_output import (
 )
 
 
-class GKInputTGLF(GKInput, FileReader, file_type="TGLF", reads=GKInput):
+class GKInputTGLF(GKInputFlat, FileReader, file_type="TGLF", reads=GKInput):
     """Reader for TGLF input files"""
 
     code_name = "TGLF"